+++ CHANGELOG.md
@@ -4,6 +4,11 @@ All notable changes to the "swagger-explorer" extension will be documented in th
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
+## [0.0.8] - 2020-01-16
+
+- Add new source button
+- Multiple config file in the same project now have different tree items
+
 ## [0.0.7] - 2020-01-13
 
 - [Swagger json are now cached](https://github.com/dardino/vscode-swagger-explorer/pull/14)